might fix the windows build, might break the jpeg lib :)

+++ lib_jpeg/jmorecfg.h
@@ -159,7 +159,7 @@ typedef short INT16;
 
 #ifndef XMD_H			/* X11/xmd.h correctly defines INT32 */
 #ifndef _BASETSD_H
-typedef long INT32;
+typedef int INT32;
 #endif
 #endif
